Unity update to 2022.3.59f1. Many status effect changes, plus sight, ruwach, and stone curse.

+++ RoRebuildServer/RoRebuildServer/Simulation/StatusEffects/Setup/StatusEffectPoolManager.cs
@@ -6,7 +6,7 @@ namespace RoRebuildServer.Simulation.StatusEffects.Setup;
 
 public struct PendingStatusEffect : IEquatable<PendingStatusEffect>
 {
-    public float EffectiveTime;
+    public double EffectiveTime;
     public bool OverwriteExisting;
     public StatusEffectState Effect;
